resinwatch-api/requirements.txt
fastapi>=0.78.0
genshin>=1.2.1
pydantic>=1.9.1
python-dotenv>=0.20.0
starlette>=0.14.2
uvicorn>=0.15.0
